requirements.txt
python-telegram-bot==13.5
git+https://github.com/elshaio/shell_pipes.git@v0.1.1
